+++ example/lib/src/loopback.dart
@@ -91,6 +91,10 @@ class _LoopbackState extends State<Loopback> {
   // Platform messages are asynchronous, so we initialize in an async method.
   void _makeCall() async {
     await setLogLevel(LogLevel.info);
+    // Correct place for testing these APIs.
+    //await setupAudioSessionManagement(false);
+    //await setOutputAudioId("speaker");
+
     var caps = DeviceConstraints();
     caps.audio.mandatory = AudioConstraints();
     caps.video.mandatory = DeviceVideoConstraints();